// src/store/products/mutation-types.ts
export enum ProductsMutationTypes {
    SET_PRODUCTS = 'SET_PRODUCTS',
    SET_PRODUCT_ID = 'SET_PRODUCT_ID',
    SET_PRODUCT='SET_PRODUCT',
    REMOVE_PRODUCT='REMOVE_PRODUCT',
    CHANGE_PRODUCTS_PAGE = 'CHANGE_PRODUCTS_PAGE',
    CHANGE_PRODUCTS_LOADING_LIST = 'CHANGE_PRODUCTS_LOADING_LIST',
    CHANGE_PRODUCTS_LOADING_ITEM = 'CHANGE_PRODUCTS_LOADING_ITEM',
    SET_PRODUCTS_COUNT='SET_PRODUCTS_COUNT'
}